training/requirements.txt
# Module 2 - QLoRA Fine-Tuning Dependencies
# Install with: pip install -r training/requirements.txt

peft>=0.10.0
transformers>=4.40.0
bitsandbytes>=0.43.0
accelerate>=0.29.0
datasets>=2.18.0
trl>=0.8.6
pyyaml>=6.0
torch>=2.1.0         # Should already be installed via vLLM
scipy>=1.11.0        # Required by some peft versions
